+++ src/main/java/cart/dao/ProductDao.java
@@ -42,11 +42,13 @@ public class ProductDao {
     public void modify(final ProductEntity modifiedProductEntity) {
         final String sql = "UPDATE product SET name = ?, price = ?, image_url = ? WHERE id = ?";
 
-        jdbcTemplate.update(sql,
-                            modifiedProductEntity.getName(),
-                            modifiedProductEntity.getPrice(),
-                            modifiedProductEntity.getImageUrl(),
-                            modifiedProductEntity.getId());
+        jdbcTemplate.update(
+                sql,
+                modifiedProductEntity.getName(),
+                modifiedProductEntity.getPrice(),
+                modifiedProductEntity.getImageUrl(),
+                modifiedProductEntity.getId()
+        );
     }
 
     public void deleteById(final Long productId) {
